test: use moto to mock SQS
