+++ .travis.yml
@@ -3,21 +3,20 @@ python:
 - '2.7'
 - '3.5'
 - '3.6'
-- '3.7-dev'
+- '3.7'
+- '3.8'
 dist: trusty
 before_install:
 - PYTHON_MAJOR_VERSION=`echo $TRAVIS_PYTHON_VERSION | head -c 1`
 - if [[ $PYTHON_MAJOR_VERSION == 3 ]]; then sudo apt-get install python3-pip; fi
-- mkdir -p ~/.cloudvolume/secrets/
-- echo $AWS_SECRET > ~/.cloudvolume/secrets/aws-secret.json
 install:
 - if [[ $PYTHON_MAJOR_VERSION == 2 ]]; then virtualenv venv; fi
 - if [[ $PYTHON_MAJOR_VERSION == 3 ]]; then virtualenv -p python3 venv; fi
 - source venv/bin/activate
-- if [[ $PYTHON_MAJOR_VERSION == 2 ]]; then pip install numpy; fi
-- if [[ $PYTHON_MAJOR_VERSION == 3 ]]; then pip3 install numpy; fi
+- if [[ $PYTHON_MAJOR_VERSION == 2 ]]; then pip install -r requirements.txt -r requirements_dev.txt; fi
+- if [[ $PYTHON_MAJOR_VERSION == 3 ]]; then pip3 install -r requirements.txt -r requirements_dev.txt; fi
 - if [[ $PYTHON_MAJOR_VERSION == 2 ]]; then pip install -e .; fi
 - if [[ $PYTHON_MAJOR_VERSION == 3 ]]; then pip3 install -e .; fi
 script:
-- if [[ $PYTHON_MAJOR_VERSION == 2 ]]; then py.test -v -x test; fi
+- if [[ $PYTHON_MAJOR_VERSION == 2 ]]; then python -m pytest -v -x test; fi
 - if [[ $PYTHON_MAJOR_VERSION == 3 ]]; then python3 -m pytest -v -x test; fi
