+++ requirements_dev.txt
@@ -1,0 +1,2 @@
+moto
+pytest
